Create VPN_status_check
I am trying to write a shell script to check the status of my VPN connections. The logic I want is If the VPN is down, it should bring up the tunnel. If that failed to do so it should send an email with relevant details.

The issue I am facing here while executing the script my else statement "ipsec up $i" being executed even for the tunnels which is up and running. 

Contain "/root/vpn_list" is my VPN Tunnel name.